adapters/requirements.txt
httpx>=0.25.0
paho-mqtt>=1.6.0
sgp4>=2.22
